lib/service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0

sqlalchemy==2.0.36
pymysql==1.1.1

python-dotenv==1.0.1

python-jose[cryptography]==3.3.0

passlib==1.7.4
argon2-cffi==23.1.0

pydantic==2.10.3
email-validator==2.2.0
typing-extensions==4.11.0